Update last libft

+++ ft_atoi_base.c
@@ -5,8 +5,8 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: fpasquer <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2016/05/31 19:30:16 by fpasquer          #+#    #+#             */
-/*   Updated: 2016/05/31 19:33:00 by fpasquer         ###   ########.fr       */
+/*   Created: 2017/01/27 21:48:57 by fpasquer          #+#    #+#             */
+/*   Updated: 2017/01/27 21:48:59 by fpasquer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,8 +36,7 @@ unsigned long long int		ret_val(char c, size_t exp, int base)
 	return (c * power);
 }
 
-int							ft_atoi_base(const char *s, int base,
-		unsigned long long int *nb)
+int							ft_atoi_base(const char *s, int base, int *nb)
 {
 	size_t					len;
 	size_t					i;
